+++ packages/core/src/hid/index.ts
@@ -1,4 +1,5 @@
 export * from "./hid-buttons";
 export * from "./hid-gamepad";
 export * from "./playstation";
+export * from "./switch";
 export * from "./xinput";